tech(api): move checkUserIsCandidate security pre-handler in certification/enrolment
